Remove InputLoopService
